MLProjects/creditcardfrauddetection/requirements.txt
# API Framework
fastapi>=0.103.2
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Security and Authentication
python-jose>=3.3.0
cryptography>=41.0.4  # Required by python-jose

# ML and Data Science
numpy>=1.26.0
pandas>=2.1.1
scikit-learn>=1.3.1
xgboost>=2.0.0
torch>=2.0.1
joblib>=1.3.2

# LLM and Vector Database
langchain>=0.0.310
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.2.0
tiktoken>=0.5.1
sentence-transformers>=2.2.2

# Vector Databases (choose one or both)
pinecone>=2.2.4  # Updated from pinecone-client to pinecone
chromadb>=0.4.15

# Utilities
tenacity>=8.2.3
loguru>=0.7.2
httpx>=0.25.0
ujson>=5.8.0
tqdm>=4.66.1
requests>=2.31.0  # For API calls to online Ollama services

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Monitoring
prometheus-client>=0.16.0
prometheus-fastapi-instrumentator>=6.0.0

# UI
streamlit>=1.31.0  # Updated to latest version
plotly>=5.18.0    # Updated to latest version
python-dotenv>=1.0.0
plotly>=5.17.0
pandas-profiling>=3.6.6
streamlit-pandas-profiling>=0.1.3

# Optional: Development tools
black>=23.9.1
isort>=5.12.0
mypy>=1.5.1
flake8>=6.1.0
pre-commit>=3.4.0
